requirements.txt
plotly
scikit-learn
pandas
umap-learn
seaborn